Fix: detalles Fix: database comprobantes

+++ database/migrations/2023_07_15_235718_create_comprobantes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
                   ->on('detalles')
                   ->onDelete("cascade")
                   ->onUpdate("cascade");
-            $table->double('total');
+            $table->float('total');
             $table->timestamp('fecha');
         });
     }
